refactor: 페이지 구조 및 타입 분리 개선
