+++ week4/assignment/src/pages/LoginPage/interfaces/LoginPage.interface.ts
@@ -1,0 +1,4 @@
+export interface LoginPageProps {
+  onLoginSuccess: (userId: string) => void;
+  onNavigateSignup: () => void;
+}
