+++ week4/assignment/src/pages/Mypage/Mypage.styles.ts
@@ -6,18 +6,18 @@ export const containerStyle = (theme: Theme) => css`
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  width: 100vw;
-  height: 100vh;
-  gap: 1rem;
+  width: ${theme.sizes.fullWidth};
+  height: ${theme.sizes.fullHeight};
+  gap: ${theme.spacing.medium};
   background-color: ${theme.colors.background};
 `;
 
 export const titleStyle = (theme: Theme) => css`
-  font-size: 2rem;
+  font-size: ${theme.fontSizes.title};
   color: ${theme.colors.primary};
 `;
 
 export const messageStyle = (theme: Theme) => css`
-  font-size: 1.2rem;
+  font-size: ${theme.fontSizes.message};
   color: ${theme.colors.text};
 `;
