+++ week4/assignment/src/pages/Mypage/interfaces/Mypage.interface.ts
@@ -1,0 +1,3 @@
+export interface MypageProps {
+  userId: string;
+}
