+++ week4/assignment/src/pages/SignupPage/interfaces/SignupPage.interface.ts
@@ -1,0 +1,3 @@
+export interface SignupPageProps {
+  onBack: () => void;
+}
