+++ week4/assignment/src/pages/SignupPage/types/SignupStep.enum.ts
@@ -1,0 +1,4 @@
+export enum SignupStep {
+  EnterId = 0,
+  EnterPassword = 1,
+}
